requirements.txt
elevenlabs==0.2.26
python-dotenv==1.0.0
moviepy==1.0.3
python-ffmpeg>=2.0.0
ffmpeg-python>=0.2.0